Text-based vampire hunting game controlled using the wasd keys.
The user plays the game through the Dungeon class, where they can create an n x m map
with an arbitrary number of vampires. The player has a limited number of moves and
must eliminate the vampires by encountering them on the map. The vampires may move as
well.

Further modifications to be made at a later time:

 - Limits on map size, number of moves, and number of vampires so that they cannot be arbitrarily large nor less than 0

 - A text-based introduction to the game explaining the rules.

 - Get user input to create the game based on the input.

 - Modify the game to include graphics.