A File Manager class with methods to both read and write to files.
Each of the methods do not handle possible errors, but instead delegate the responsibility
to the method caller. The read method takes the name of a file as a string and returns an
ArrayList of the lines within the file. The two save methods also take the name of a file
and either write a single string or multiple strings to a file. If the file already exists,
it is overwritten.